apps/python/requirements.txt
azure-functions
azure-monitor-opentelemetry
requests
